adding contains, is in (+ not) grammar and semantics + tests
note: these conditional grammar rules accept any expressions, but
only strings are tested
