+++ js/ohm/simpletalk.ohm
@@ -104,6 +104,18 @@ SimpleTalk {
       = Expression "is not a" Expression
       | Expression "is not an" Expression
 
+    ContainsConditional
+      = Expression "contains" Expression
+
+    DoesNotContainConditional
+      = Expression "does not contain" Expression
+
+    IsInConditional
+      = Expression "is in" Expression
+    
+    IsNotInConditional
+      = Expression "is not in" Expression
+    
     ThereIsAnObjectConditional
       = "there" "is" ("an" | "a") ObjectSpecifier
 
@@ -120,6 +132,10 @@ SimpleTalk {
 
     Conditional
       = booleanLiteral --booleanLiteral
+      | IsInConditional
+      | IsNotInConditional
+      | ContainsConditional
+      | DoesNotContainConditional
       | ThereIsAnObjectConditional
       | ThereIsNotAnObjectConditional
       | ThereIsAPropertyConditional
